+++ CHANGELOG.md
@@ -31,6 +31,9 @@ Enter the dimension by throwing an enderpearl at Bee Nest and exit it by going a
   
 - Bee Dungeon and Spider Infested Bee Dungeon code backend is significantly cleaned up and now uses processors to change blocks.
 
+   Mod Compat:
+- Added Charm support so now their candles spawn in Bee Dungeons and Spider Infested Bee Dungeons!
+
 
 
    (V.2.3.0 Changes) (1.16.5 Minecraft)